requirements.txt
asyncpg==0.29.0
fastapi==0.109.0
Flask==3.0.1
passlib==1.7.4
psycopg2==2.9.9
pydantic==2.5.3
PyJWT==2.7.0
PyJWT==2.8.0
PyJWT==2.8.0
starlette==0.35.1
